requirements.txt
khl.py==0.3.2
